requirements.txt
aiofiles
pyyaml
telethon
tgcrypto
